Refactor imports to utilize FlowErrorPageContent from core/ui for consistent error handling across components. This change enhances the error display mechanism in various parts of the application.
