+++ clients/desktop/src/vault/import/components/ReadBackupFileStep.tsx
@@ -1,13 +1,11 @@
 import { useAppViewState } from '@clients/desktop/src/navigation/hooks/useAppViewState'
 import { ReadTextFile } from '@clients/desktop/wailsjs/go/main/App'
+import { FlowErrorPageContent } from '@core/ui/flow/FlowErrorPageContent'
 import { isLikelyToBeDklsVaultBackup } from '@core/ui/vault/import/utils/isLikelyToBeDklsVaultBackup'
 import { vaultContainerFromString } from '@core/ui/vault/import/utils/vaultContainerFromString'
 import { FileBasedVaultBackupResult } from '@core/ui/vault/import/VaultBackupResult'
-import { Button } from '@lib/ui/buttons/Button'
-import { FlowErrorPageContent } from '@lib/ui/flow/FlowErrorPageContent'
 import { FlowPageHeader } from '@lib/ui/flow/FlowPageHeader'
 import { FlowPendingPageContent } from '@lib/ui/flow/FlowPendingPageContent'
-import { useNavigateBack } from '@lib/ui/navigation/hooks/useNavigateBack'
 import { OnFinishProp } from '@lib/ui/props'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
 import { useMutation } from '@tanstack/react-query'
@@ -19,7 +17,6 @@ export const ReadBackupFileStep = ({
 }: OnFinishProp<FileBasedVaultBackupResult>) => {
   const { t } = useTranslation()
   const [{ filePath }] = useAppViewState<'importVaultFromFile'>()
-  const navigateBack = useNavigateBack()
 
   const { mutate, ...mutationState } = useMutation({
     mutationFn: async () => {
@@ -57,7 +54,6 @@ export const ReadBackupFileStep = ({
         pending={() => <FlowPendingPageContent title={t('importing_vault')} />}
         error={error => (
           <FlowErrorPageContent
-            action={<Button onClick={navigateBack}>{t('back')}</Button>}
             title={t('failed_to_import_vault')}
             error={error}
           />
