+++ core/ui/storage/StorageDependant.tsx
@@ -1,11 +1,11 @@
 import { ProductLogoBlock } from '@core/ui/product/ProductLogoBlock'
-import { FlowErrorPageContent } from '@lib/ui/flow/FlowErrorPageContent'
 import { NavigationProvider } from '@lib/ui/navigation/state'
 import { ChildrenProp } from '@lib/ui/props'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
 import { useMergeQueries } from '@lib/ui/query/hooks/useMergeQueries'
 
 import { RootErrorBoundary } from '../errors/RootErrorBoundary'
+import { FlowErrorPageContent } from '../flow/FlowErrorPageContent'
 import { I18nProvider } from '../i18n/I18nProvider'
 import { PasscodeProvider } from '../passcodeEncryption/state/passcode'
 import { RootCurrentVaultProvider } from '../vault/state/currentVault'
