+++ core/ui/vault/save/SaveVaultStep.tsx
@@ -1,13 +1,14 @@
 import { useCreateVaultMutation } from '@core/ui/vault/mutations/useCreateVaultMutation'
 import { Vault } from '@core/ui/vault/Vault'
 import { Button } from '@lib/ui/buttons/Button'
-import { FlowErrorPageContent } from '@lib/ui/flow/FlowErrorPageContent'
 import { FlowPageHeader } from '@lib/ui/flow/FlowPageHeader'
 import { FlowPendingPageContent } from '@lib/ui/flow/FlowPendingPageContent'
 import { OnBackProp, OnFinishProp, TitleProp, ValueProp } from '@lib/ui/props'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
+
+import { FlowErrorPageContent } from '../../flow/FlowErrorPageContent'
 
 export const SaveVaultStep: React.FC<
   ValueProp<Vault> & OnFinishProp & TitleProp & OnBackProp
